requirements.txt
openai>=1.30.0
python-telegram-bot>=20.8
